third_party/WebKit/LayoutTests/editing/selection/collapseto_in_text_fields.html

<!DOCTYPE html>
<title>Tests for crbug.com/716725</title>
<body>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>

<input value="abcdef">
<textarea>abcdef</textarea>

<script>
function testCollapseToStart(element) {
  test(() => {
    element.focus();
    element.setSelectionRange(1, 5);
    getSelection().collapseToStart();
    assert_equals(element.selectionStart, 1);
    assert_equals(element.selectionEnd, 1);
  }, 'Selection.collapseToStart() should collapse text selection in ' + element.tagName);
}

function testCollapseToEnd(element) {
  test(() => {
    element.focus();
    element.setSelectionRange(1, 5);
    getSelection().collapseToEnd();
    assert_equals(element.selectionStart, 5);
    assert_equals(element.selectionEnd, 5);
  }, 'Selection.collapseToEnd() should collapse text selection in ' + element.tagName);
}

testCollapseToStart(document.querySelector('input'));
testCollapseToStart(document.querySelector('textarea'));
testCollapseToEnd(document.querySelector('input'));
testCollapseToEnd(document.querySelector('textarea'));
</script>
</body>
